# rawg-server/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Game": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description_raw": {
						"type": "string"
					},
					"metacritic": {
						"type": "number",
						"format": "double"
					},
					"background_image": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"released": {
						"type": "string"
					},
					"added": {
						"type": "number",
						"format": "double"
					},
					"rating_top": {
						"type": "number",
						"format": "double"
					},
					"website": {
						"type": "string"
					},
					"genres": {
						"items": {
							"$ref": "#/components/schemas/Genre"
						},
						"type": "array"
					},
					"parent_platforms": {
						"items": {
							"$ref": "#/components/schemas/ParentPlatform"
						},
						"type": "array"
					},
					"stores": {
						"items": {
							"$ref": "#/components/schemas/Store"
						},
						"type": "array"
					},
					"publishers": {
						"items": {
							"$ref": "#/components/schemas/Publisher"
						},
						"type": "array"
					},
					"developers": {
						"items": {
							"$ref": "#/components/schemas/Developer"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array"
					},
					"wishlistedBy": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"inLibraryOf": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"trailers": {
						"items": {
							"$ref": "#/components/schemas/Trailer"
						},
						"type": "array"
					},
					"screenshots": {
						"items": {
							"$ref": "#/components/schemas/Screenshot"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"genres",
					"parent_platforms",
					"stores",
					"publishers",
					"developers",
					"tags",
					"wishlistedBy",
					"inLibraryOf",
					"trailers",
					"screenshots"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Genre": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParentPlatform": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Store": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Publisher": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Developer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tag": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"language",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"user"
						]
					},
					"wishlist": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					},
					"library": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"username",
					"passwordHash",
					"role",
					"wishlist",
					"library"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Trailer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"preview": {
						"type": "string"
					},
					"data480": {
						"type": "string"
					},
					"dataMax": {
						"type": "string"
					},
					"game": {
						"$ref": "#/components/schemas/Game"
					}
				},
				"required": [
					"id",
					"name",
					"preview",
					"data480",
					"dataMax",
					"game"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Screenshot": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"is_deleted": {
						"type": "boolean"
					},
					"game": {
						"$ref": "#/components/schemas/Game"
					}
				},
				"required": [
					"id",
					"image",
					"width",
					"height",
					"is_deleted",
					"game"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ListResponse_Genre_": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/Genre"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Genre_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_Game_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description_raw": {
						"type": "string"
					},
					"metacritic": {
						"type": "number",
						"format": "double"
					},
					"background_image": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"released": {
						"type": "string"
					},
					"added": {
						"type": "number",
						"format": "double"
					},
					"rating_top": {
						"type": "number",
						"format": "double"
					},
					"website": {
						"type": "string"
					},
					"genres": {
						"items": {
							"$ref": "#/components/schemas/Genre"
						},
						"type": "array"
					},
					"parent_platforms": {
						"items": {
							"$ref": "#/components/schemas/ParentPlatform"
						},
						"type": "array"
					},
					"stores": {
						"items": {
							"$ref": "#/components/schemas/Store"
						},
						"type": "array"
					},
					"publishers": {
						"items": {
							"$ref": "#/components/schemas/Publisher"
						},
						"type": "array"
					},
					"developers": {
						"items": {
							"$ref": "#/components/schemas/Developer"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array"
					},
					"wishlistedBy": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"inLibraryOf": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"trailers": {
						"items": {
							"$ref": "#/components/schemas/Trailer"
						},
						"type": "array"
					},
					"screenshots": {
						"items": {
							"$ref": "#/components/schemas/Screenshot"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ListResponse_Developer_": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/Developer"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Developer_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image_background": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rawg-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/genres": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_Genre_"
								}
							}
						}
					}
				},
				"tags": [
					"Genres"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Genre"
								}
							}
						}
					}
				},
				"tags": [
					"Genres"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Genre_"
							}
						}
					}
				}
			}
		},
		"/genres/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Genre"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Genres"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Genre"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Genres"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Genre_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Genres"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"items": {
												"properties": {
													"screenshots": {
														"items": {
															"$ref": "#/components/schemas/Screenshot"
														},
														"type": "array"
													},
													"trailers": {
														"items": {
															"$ref": "#/components/schemas/Trailer"
														},
														"type": "array"
													},
													"tags": {
														"items": {
															"$ref": "#/components/schemas/Tag"
														},
														"type": "array"
													},
													"developers": {
														"items": {
															"$ref": "#/components/schemas/Developer"
														},
														"type": "array"
													},
													"publishers": {
														"items": {
															"$ref": "#/components/schemas/Publisher"
														},
														"type": "array"
													},
													"stores": {
														"items": {
															"$ref": "#/components/schemas/Store"
														},
														"type": "array"
													},
													"genres": {
														"items": {
															"$ref": "#/components/schemas/Genre"
														},
														"type": "array"
													},
													"website": {
														"type": "string"
													},
													"rating_top": {
														"type": "number",
														"format": "double"
													},
													"added": {
														"type": "number",
														"format": "double"
													},
													"released": {
														"type": "string"
													},
													"rating": {
														"type": "number",
														"format": "double"
													},
													"background_image": {
														"type": "string"
													},
													"metacritic": {
														"type": "number",
														"format": "double"
													},
													"description_raw": {
														"type": "string"
													},
													"slug": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													},
													"inLibraryOf": {
														"items": {
															"properties": {
																"username": {
																	"type": "string"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"username",
																"id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"wishlistedBy": {
														"items": {
															"properties": {
																"username": {
																	"type": "string"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"username",
																"id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"parent_platforms": {
														"items": {
															"properties": {
																"platform": {
																	"$ref": "#/components/schemas/ParentPlatform"
																}
															},
															"required": [
																"platform"
															],
															"type": "object"
														},
														"type": "array"
													}
												},
												"required": [
													"screenshots",
													"trailers",
													"tags",
													"developers",
													"publishers",
													"stores",
													"genres",
													"slug",
													"name",
													"id",
													"inLibraryOf",
													"wishlistedBy",
													"parent_platforms"
												],
												"type": "object"
											},
											"type": "array"
										},
										"next": {
											"type": "string"
										},
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"results",
										"next",
										"count"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page_size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "genreId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "storeId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "platformId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "publisherId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "developerId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "wishlistUserId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "libraryUserId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "searchText",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tagId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Game"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Game_"
							}
						}
					}
				}
			}
		},
		"/games/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Game"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Game"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Game_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/{id}/movies": {
			"get": {
				"operationId": "GetTrailers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"items": {
												"$ref": "#/components/schemas/Trailer"
											},
											"type": "array"
										},
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"results",
										"count"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/{id}/screenshots": {
			"get": {
				"operationId": "GetScreenshots",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"results": {
											"items": {
												"$ref": "#/components/schemas/Screenshot"
											},
											"type": "array"
										},
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"results",
										"count"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/developers": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_Developer_"
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Developer"
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Developer_"
							}
						}
					}
				}
			}
		},
		"/developers/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Developer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Developer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Developer_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Developers"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}